+++ src/conversion/parseCommand.test.ts
@@ -251,14 +251,14 @@ describe('how parseCommand works', () => {
     });
 
     it('should parse EXPECT command', () => {
-        expect(parseCommand('Expect 1 character')).toEqual({
+        expect(parseCommand('Expect exactly 1 character')).toEqual({
             type: 'EXPECT',
             sign: 'EXACTLY',
             unit: 'character',
             amount: 1,
         });
 
-        expect(parseCommand('Expect 1 char')).toEqual({
+        expect(parseCommand('Expect exactly 1 char')).toEqual({
             type: 'EXPECT',
             sign: 'EXACTLY',
             unit: 'character',
@@ -293,76 +293,98 @@ describe('how parseCommand works', () => {
             amount: 5,
         });
 
-        expect(parseCommand('Expect maximally 5 character')).toEqual({
+        expect(parseCommand('Expect maximally 5 characters')).toEqual({
             type: 'EXPECT',
             sign: 'MAXIMUM',
             unit: 'character',
             amount: 5,
         });
 
-        expect(parseCommand('Expect max 5 char')).toEqual({
+        expect(parseCommand('Expect max 5 CHARs')).toEqual({
             type: 'EXPECT',
             sign: 'MAXIMUM',
             unit: 'character',
             amount: 5,
         });
 
-        expect(parseCommand('Expect 1 word')).toEqual({
+        expect(parseCommand('Expect exact 1 word')).toEqual({
             type: 'EXPECT',
             sign: 'EXACTLY',
             unit: 'word',
             amount: 1,
         });
-        expect(parseCommand('Expect 2 words')).toEqual({
+
+        expect(parseCommand('Expect exactly 1 word')).toEqual({
+            type: 'EXPECT',
+            sign: 'EXACTLY',
+            unit: 'word',
+            amount: 1,
+        });
+
+        expect(parseCommand('Expect eXactly 1 word')).toEqual({
+            type: 'EXPECT',
+            sign: 'EXACTLY',
+            unit: 'word',
+            amount: 1,
+        });
+
+        expect(parseCommand('EXPECT EXACTLY 1 WORD')).toEqual({
+            type: 'EXPECT',
+            sign: 'EXACTLY',
+            unit: 'word',
+            amount: 1,
+        });
+
+        expect(parseCommand('Expect exactly 2 words')).toEqual({
             type: 'EXPECT',
             sign: 'EXACTLY',
             unit: 'word',
             amount: 2,
         });
 
-        expect(parseCommand('Expect 1 sentence')).toEqual({
+        expect(parseCommand('Expect exactly 1 sentence')).toEqual({
             type: 'EXPECT',
             sign: 'EXACTLY',
             unit: 'sentence',
             amount: 1,
         });
 
-        expect(parseCommand('Expect 2 sentences')).toEqual({
+        expect(parseCommand('Expect exactly 2 sentences')).toEqual({
             type: 'EXPECT',
             sign: 'EXACTLY',
             unit: 'sentence',
             amount: 2,
         });
 
-        expect(parseCommand('Expect 0 sentences')).toEqual({
+        expect(parseCommand('Expect exactly 0 sentences')).toEqual({
             type: 'EXPECT',
             sign: 'EXACTLY',
             unit: 'sentence',
             amount: 0,
         });
 
-        expect(parseCommand('Expect 1 paragraph')).toEqual({
+        expect(parseCommand('Expect exactly 1 paragraph')).toEqual({
             type: 'EXPECT',
             sign: 'EXACTLY',
             unit: 'paragraph',
             amount: 1,
         });
 
-        expect(parseCommand('Expect 2 paragraphs')).toEqual({
+        expect(parseCommand('Expect exactly 2 paragraphs')).toEqual({
             type: 'EXPECT',
             sign: 'EXACTLY',
             unit: 'paragraph',
             amount: 2,
         });
 
-        expect(parseCommand('Expect 1 line')).toEqual({
+        expect(parseCommand('Expect exactly 1 line')).toEqual({
             type: 'EXPECT',
             sign: 'EXACTLY',
             unit: 'line',
             amount: 1,
         });
 
-        expect(parseCommand('Expect 2 lines')).toEqual({
+        expect(parseCommand('Expect exactly 2 lines')).toEqual({
             type: 'EXPECT',
             sign: 'EXACTLY',
             unit: 'line',
@@ -397,6 +419,13 @@ describe('how parseCommand works', () => {
             /Can not contain another parameter in description/i,
         );
         expect(() => parseCommand('parmeter {name} Name for the hero')).toThrowError(/Unknown command/i);
+    });
+
+    it('should fail parsing expect command', () => {
+        expect(() => parseCommand('Expect foo 1 char')).toThrowError(/Invalid EXPECT command/i);
+        expect(() => parseCommand('Expect min 1 vars')).toThrowError(/Invalid EXPECT command/i);
+        expect(() => parseCommand('Expect min chars')).toThrowError(/Invalid EXPECT command/i);
+        expect(() => parseCommand('Expect min xx chars')).toThrowError(/Invalid EXPECT command/i);
     });
 
     it('should fail parsing multiline command', () => {