+++ src/utils/expectation-counters/countLines.ts
@@ -1,9 +1,9 @@
-import type { number_integer, number_positive } from '../../types/typeAliases';
+import type { ExpectationAmount } from '../../types/PromptTemplatePipelineJson/PromptTemplateJson';
 
 /**
  * Counts mumber of lines in the text
  */
 
-export function countLines(text: string): number_integer & number_positive {
+export function countLines(text: string): ExpectationAmount{
     return 0; /* <- !!!!! Implement */
 }