+++ src/utils/expectation-counters/countPages.ts
@@ -1,9 +1,9 @@
-import type { number_integer, number_positive } from '../../types/typeAliases';
+import type { ExpectationAmount } from '../../types/PromptTemplatePipelineJson/PromptTemplateJson';
 
 /**
  * Counts mumber of pages in the text
  */
 
-export function countPages(text: string): number_integer & number_positive {
+export function countPages(text: string):ExpectationAmount{
     return 0; /* <- !!!!! Implement */
 }