+++ src/utils/expectation-counters/countWords.ts
@@ -1,9 +1,9 @@
-import type { number_integer, number_positive } from '../../types/typeAliases';
+import type { ExpectationAmount } from '../../types/PromptTemplatePipelineJson/PromptTemplateJson';
 
 /**
  * Counts mumber of words in the text
  */
 
-export function countWords(text: string): number_integer & number_positive {
+export function countWords(text: string): ExpectationAmount {
     return 0; /* <- !!!!! Implement */
 }